Add DDS subscriber layer, game/chat publishers and thread manager
- Introduce dedicated Subscriber classes (Chat/Game/InRoom/JoinRoom/
  PlayerAuth/PlayerColorMappings) under data/dds/subscriber
- Add ChatPublisher and GamePublisher under data/dds/publisher
- Add ThreadManager to centrally own and route subscriber threads
- Rework GameSurfaceView rendering and MainActivity game flow
- Ignore local session dir and java_tools build output

+++ app/src/main/java/com/example/snakegame/presentation/contract/GameContract.java
@@ -26,8 +26,8 @@ public interface GameContract {
     interface Presenter {
         void attachView(View view);
         void detachView();
-        void initializeGame(long playerId, String nickname, String color);
-        void initializeTimedScoreMode(long playerId, String nickname, String color);
+        void initializeGame(int playerId, String nickname, String color);
+        void initializeTimedScoreMode(int playerId, String nickname, String color);
         void handlePlayerMove(String direction);
         void sendChatMessage(String message);
         void startGame();
